% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colsummarize.R
\name{colsummarize}
\alias{colsummarize}
\title{Column Summary}
\usage{
colsummarize(dfi)
}
\arguments{
\item{dfi}{data.frame containing only numerical values}
}
\description{
This function calculates the mean, sd and n of each column
}
\examples{
library("Sleuth3")
data("case0202")
colsummarize(case0202)
}
\keyword{jcpackr}

